+++ soyutnet/constants.py
@@ -1,3 +1,5 @@
+import random
+import string
 import weakref
 from weakref import ReferenceType
 from typing import (
@@ -30,6 +32,13 @@ GENERIC_ID: id_t = 0
 INITIAL_ID: id_t = 0
 
 
+def random_identifier(N: int = 5) -> str:
+    return "".join(
+        random.SystemRandom().choice(string.ascii_uppercase + string.digits)
+        for _ in range(N)
+    )
+
+
 class BaseObject(object):
     """
     Base SoyutNet object inherited by all classes.
@@ -37,15 +46,26 @@ class BaseObject(object):
 
     def __init__(self, net: "SoyutNet") -> None:
         self._net: ReferenceType["SoyutNet"] = weakref.ref(net)
-        self._ident: str = ""
-        """Every object should have a unique string identifier."""
+        """Reference to the creator SoyutNet instance."""
+        self._ident0: str = random_identifier()
 
     @property
     def net(self) -> "SoyutNet":
         """
-        Reference of the SoyutNet instance assigned to instances of all object types.
+        Get reference of the SoyutNet instance assigned to instances of all object types,
+        for a particular simulation.
+
+        :return: Reference to the creator SoyutNet instance.
         """
         return self._net()  # type: ignore[return-value]
+
+    def ident(self) -> str:
+        """
+        Get object's unique identifier
+
+        :return: Identifier string.
+        """
+        return self._ident0
 
 
 class SoyutNetError(Exception):