+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# Version 0.3.0
+
+- Fix mistakes in PT net implementation
+- Improved documentation
+- Wrapped `asyncio.run` in a clean interface
+  - It also supports attaching asyncio tasks additional to the PT loops
+- Improved graphviz generator
+  - Custom indentation and label names
+  - Custom coloring for `SpecialPlace`
+- Can subscribe to transition firing notifications from any task
+- Custom formatted debug printing
+- Added new tests and examples
+
+
 ## Version 0.2.0
 
 - Improved documentation